i just put all the challenges

+++ aymane_app/ch4.c
@@ -1,15 +1,15 @@
 #include<stdio.h>
 
 int main (){
-    int a,b,c,d,sum,avr;
+    float a,b,c,d,sum,avr;
 
     printf("enter the four numbers :: ");
-    scanf("%d %d %d %d", &a,&b,&c,&d);
+    scanf("%f %f %f %f", &a,&b,&c,&d);
 
     sum = a+b+c+d;
     avr = (a+b+c+d)/4;
-    printf("the sum is :: %d\n", sum);
-    printf("the avr is :: %d",avr);
+    printf("the sum is :: %f\n", sum);
+    printf("the avr is :: %f",avr);
 
 
     return 0;
